Major Mod (Dylan Torres)
Implemented the major mod: a mini-map that displays the entire level, achieved through including a second camera. Expanded the level fog in the level prefab to make presentation more consistent, and added a tracking spotlight to John Lemon's feet in the John Lemon prefab to enhance his position's readability on the mini-map.